+++ src/quantum/azext_quantum/vendored_sdks/azure_quantum_python/_client/operations/_operations.py
@@ -332,6 +332,35 @@ def build_services_quotas_list_request(
     return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
 
 
+def build_services_quotas_list_quota_usages_request(  # pylint: disable=name-too-long
+    subscription_id: str, resource_group_name: str, workspace_name: str, provider_id: str, **kwargs: Any
+) -> HttpRequest:
+    _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
+    _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
+
+    api_version: str = kwargs.pop("api_version", _params.pop("api-version", "2026-01-15-preview"))
+    accept = _headers.pop("Accept", "application/json")
+
+    # Construct URL
+    _url = "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Quantum/workspaces/{workspaceName}/quotaUsages"
+    path_format_arguments = {
+        "subscriptionId": _SERIALIZER.url("subscription_id", subscription_id, "str"),
+        "resourceGroupName": _SERIALIZER.url("resource_group_name", resource_group_name, "str"),
+        "workspaceName": _SERIALIZER.url("workspace_name", workspace_name, "str"),
+    }
+
+    _url: str = _url.format(**path_format_arguments)  # type: ignore
+
+    # Construct parameters
+    _params["providerId"] = _SERIALIZER.query("provider_id", provider_id, "str")
+    _params["api-version"] = _SERIALIZER.query("api_version", api_version, "str")
+
+    # Construct headers
+    _headers["Accept"] = _SERIALIZER.header("accept", accept, "str")
+
+    return HttpRequest(method="GET", url=_url, params=_params, headers=_headers, **kwargs)
+
+
 def build_services_suite_offers_list_quota_usages_request(  # pylint: disable=name-too-long
     subscription_id: str, provider_id: str, **kwargs: Any
 ) -> HttpRequest:
@@ -1740,6 +1769,72 @@ class ServicesQuotasOperations:
             return pipeline_response
 
         return ItemPaged(get_next, extract_data)
+
+    @distributed_trace
+    def list_quota_usages(
+        self, subscription_id: str, resource_group_name: str, workspace_name: str, provider_id: str, **kwargs: Any
+    ) -> "list[_models.QuotaUsage]":
+        """List quota usages for a provider in the given workspace.
+
+        :param subscription_id: The Azure subscription ID. Required.
+        :type subscription_id: str
+        :param resource_group_name: Name of the Azure resource group. Required.
+        :type resource_group_name: str
+        :param workspace_name: Name of the Azure Quantum workspace. Required.
+        :type workspace_name: str
+        :param provider_id: The provider whose quota usages are requested. Required.
+        :type provider_id: str
+        :return: list of QuotaUsage
+        :rtype: list[~azure.quantum.models.QuotaUsage]
+        :raises ~azure.core.exceptions.HttpResponseError:
+        """
+        _headers = kwargs.pop("headers", {}) or {}
+        _params = kwargs.pop("params", {}) or {}
+
+        cls: ClsType[list[_models.QuotaUsage]] = kwargs.pop("cls", None)
+
+        error_map: MutableMapping = {
+            401: ClientAuthenticationError,
+            404: ResourceNotFoundError,
+            409: ResourceExistsError,
+            304: ResourceNotModifiedError,
+        }
+        error_map.update(kwargs.pop("error_map", {}) or {})
+
+        _request = build_services_quotas_list_quota_usages_request(
+            subscription_id=subscription_id,
+            resource_group_name=resource_group_name,
+            workspace_name=workspace_name,
+            provider_id=provider_id,
+            api_version=self._config.api_version,
+            headers=_headers,
+            params=_params,
+        )
+        path_format_arguments = {
+            "endpoint": self._serialize.url(
+                "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
+            ),
+        }
+        _request.url = self._client.format_url(_request.url, **path_format_arguments)
+
+        _stream = False
+        pipeline_response: PipelineResponse = self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
+        )
+        response = pipeline_response.http_response
+
+        if response.status_code not in [200]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            raise HttpResponseError(response=response)
+
+        deserialized = response.json()
+        # The service may return either a bare JSON array or a paged envelope ({"value": [...]}).
+        if isinstance(deserialized, dict):
+            deserialized = deserialized.get("value", [])
+        list_of_elem = _deserialize(list[_models.QuotaUsage], deserialized)
+        if cls:
+            return cls(list_of_elem)  # type: ignore
+        return list_of_elem
 
 
 class ServicesSuiteOffersOperations: